Empty webpack fixture directories
