--- tests/fixtures/simple_js/index.js
@@ -1,11 +1,0 @@
-addEventListener('fetch', event => {
-    event.respondWith(handleRequest(event.request))
-  })
-  
-  /**
-   * Fetch and log a request
-   * @param {Request} request
-   */
-  async function handleRequest(request) {
-    return new Response('Hello worker!', { status: 200 })
-  }
