--- tests/fixtures/simple_js/package.json
@@ -1,3 +1,0 @@
-{
-    "main": "index.js"
-}
