--- tests/fixtures/webpack_function_config_js/index.js
@@ -1,1 +1,0 @@
-// js
