--- tests/fixtures/webpack_function_config_js/package.json
@@ -1,1 +1,0 @@
-{}
