--- tests/fixtures/webpack_function_config_js/webpack.config.js
@@ -1,1 +1,0 @@
-module.exports = () => ({ entry: "./index.js" });
