--- tests/fixtures/webpack_function_promise_config_js/package.json
@@ -1,3 +1,0 @@
-{
-    "main": "index.js"
-}
