--- tests/fixtures/webpack_multiple_config/package.json
@@ -1,3 +1,0 @@
-{
-    "main": "index.js"
-}
