--- tests/fixtures/webpack_promise_config_js/index.js
@@ -1,1 +1,0 @@
-// js
