--- tests/fixtures/webpack_promise_config_js/package.json
@@ -1,1 +1,0 @@
-{}
