--- tests/fixtures/webpack_promise_config_js/webpack.config.js
@@ -1,1 +1,0 @@
-module.exports = Promise.resolve({ entry: "./index.js" });
