--- tests/fixtures/webpack_single_js_missing_package_main/package.json
@@ -1,3 +1,0 @@
-{
-  "name": "webpack_single_js_missing_package_main"
-}
