--- tests/fixtures/webpack_single_js_use_package_main/index.js
@@ -1,1 +1,0 @@
-// index
