--- tests/fixtures/webpack_single_js_use_package_main/package.json
@@ -1,4 +1,0 @@
-{
-  "name": "webpack_single_js_use_package_main",
-  "main": "index.js"
-}
