--- tests/fixtures/webpack_target_node/package.json
@@ -1,3 +1,0 @@
-{
-    "main": "index.js"
-}
