--- tests/fixtures/webpack_target_web/package.json
@@ -1,3 +1,0 @@
-{
-    "main": "index.js"
-}
