+++ tests/preview.rs
@@ -6,8 +6,6 @@ pub mod utils;
 use assert_cmd::prelude::*;
 
 use std::env;
-use std::fs::File;
-use std::io::Write;
 use std::process::Command;
 use std::str;
 use std::sync::Mutex;
@@ -16,31 +14,34 @@ lazy_static! {
     static ref BUILD_LOCK: Mutex<u8> = Mutex::new(0);
 }
 
-macro_rules! settings {
-    ( $f:expr, $x:expr ) => {
-        let file_path = utils::fixture_path($f).join("wrangler.toml");
-        let mut file = File::create(file_path).unwrap();
-        let content = format!(
-            r#"
-            name = "test"
-            zone_id = ""
-            account_id = ""
-            workers_dev = true
-            {}
-        "#,
-            $x
-        );
-        file.write_all(content.as_bytes()).unwrap();
-    };
-}
-
 #[test]
 fn it_can_preview_js_project() {
     let fixture = "simple_js";
     utils::create_temporary_copy(fixture);
-    settings! {fixture, r#"
+    utils::create_fixture_file(
+        fixture,
+        "index.js",
+        r#"
+        addEventListener('fetch', event => {
+            event.respondWith(handleRequest(event.request))
+        })
+        
+        /**
+        * Fetch and log a request
+        * @param {Request} request
+        */
+        async function handleRequest(request) {
+            return new Response('Hello worker!', { status: 200 })
+        }
+    "#,
+    );
+    utils::create_default_package_json(fixture);
+    utils::create_wrangler_toml(
+        fixture,
+        r#"
         type = "javascript"
-    "#};
+    "#,
+    );
     preview(fixture);
     utils::cleanup(fixture);
 }
@@ -49,9 +50,14 @@ fn it_can_preview_js_project() {
 fn it_can_preview_webpack_project() {
     let fixture = "webpack_simple_js";
     utils::create_temporary_copy(fixture);
-    settings! {fixture, r#"
+    utils::create_default_package_json(fixture);
+    utils::create_empty_js(fixture);
+    utils::create_wrangler_toml(
+        fixture,
+        r#"
         type = "webpack"
-    "#};
+    "#,
+    );
     preview(fixture);
     utils::cleanup(fixture);
 }
@@ -60,9 +66,13 @@ fn it_can_preview_webpack_project() {
 fn it_can_preview_rust_project() {
     let fixture = "simple_rust";
     utils::create_temporary_copy(fixture);
-    settings! {fixture, r#"
+    utils::create_empty_package_json(fixture);
+    utils::create_wrangler_toml(
+        fixture,
+        r#"
         type = "rust"
-    "#};
+    "#,
+    );
     preview(fixture);
     utils::cleanup(fixture);
 }
